// src/category/enums/category-message.enum.ts
export enum CategoryMessage {
    NOT_FOUND = 'Category not found',
    ALREADY_EXISTS = 'Category already exists',
    CREATED = 'Category created successfully',
    UPDATED = 'Category updated successfully',
    DELETED = 'Category deleted successfully',
    GET_CLUB_CATEGORIES = 'Get club categories successfully',
    GET_ALL_CATEGORIES = 'Get all categories successfully',
    GET_CATEGORY = 'Get category successfully',
}